Moves the ui code into its own directory.
