+++ packages/reactotron-cli/src/index.js
@@ -4,19 +4,13 @@ import Context from './context'
 import Router from './router'
 import commands from './commands/index'
 import ui from './ui'
-import uiReactions from './uiReactions'
+import reactions from './ui/reactions'
 
 const PORT = 9090
 const server = createServer({
   port: PORT,
   onCommand: command => {
     context.post(command)
-  },
-  onStart: () => {
-    context.log(`Started on port ${PORT}`)
-  },
-  onStop: () => {
-    context.log('stopped')
   },
   onConnect: client => {
     context.post({ type: 'redux.subscribe.request' })
@@ -34,7 +28,7 @@ const context = new Context({
 
 // some parts of the ui can react to mobx changes.  they go in here.  so awesome.
 // i'd love to move more in here.
-context.uiReactions = uiReactions(context)
+reactions(context)
 
 // always control-c to die
 ui.screen.key('C-c', () => context.post({type: 'program.die'}))
