--- packages/reactotron-cli/src/ui.js
@@ -1,7 +1,0 @@
-import layout from './uiLayout'
-import drawing from './uiDrawing'
-
-export default {
-  ...layout,
-  ...drawing(layout)
-}
