+++ packages/reactotron-cli/src/ui/index.js
@@ -1,0 +1,7 @@
+import layout from './layout'
+import drawing from './drawing'
+
+export default {
+  ...layout,
+  ...drawing(layout)
+}
